🖌 Move from scan.js to file.js ( same as api endpoint files)
