+++ src/middlewares/file.js
@@ -3,10 +3,10 @@ import { createToast } from 'mosha-vue-toastify';
 export default async ({ store, to, next }) => {
     let fileHash = to.params.id;
 
-    let selectedFile = await store.getters['scan/getFile'];
+    let selectedFile = await store.getters['file/getFile'];
 
     if (!selectedFile || selectedFile.sha256 != fileHash) {
-        selectedFile = store.dispatch('scan/fetchFile', fileHash);
+        selectedFile = store.dispatch('file/fetchFile', fileHash);
 
         await selectedFile.catch(err => {
             if (err) {
