+++ src/modules/file/pages/Summary.vue
@@ -64,7 +64,7 @@ import Card from "@/common/components/elements/Card.vue";
 import Progress from "@/common/components/Progress.vue";
 import TableCols from "@/common/components/tables/TableCols.vue";
 import Gallery from "@/common/components/tables/Gallery.vue";
-import { scanGetters } from "@/state/helpers";
+import { fileGetters } from "@/state/helpers";
 import { isAnAVG } from "@/common/functions";
 
 import {
@@ -88,7 +88,7 @@ export default {
     };
   },
   computed: {
-    ...scanGetters,
+    ...fileGetters,
     getProperties() {
       let items = this.file.properties || {};
