+++ src/modules/file/pages/static-analysis/Strings.vue
