+++ src/state/store.js
@@ -2,6 +2,7 @@ import Vuex from 'vuex'
 import dispatchActionForAllModules from '@/common/utils/dispatch-action-for-all-modules'
 
 import modules from './modules'
+import config from '../common/config'
 
 
 const store = new Vuex.Store({
@@ -9,7 +10,7 @@ const store = new Vuex.Store({
   // Enable strict mode in development to get a warning
   // when mutating state outside of a mutation.
   // https://vuex.vuejs.org/guide/strict.html
-  strict: process.env.NODE_ENV !== 'production',
+  strict: !config.isProd,
 })
 
 export default store
